pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lendbase"
version = "0.1.0"
description = "Internal inventory and lending app for a university admin team."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "HU Berlin admin tooling team" }
]
dependencies = [
  "SQLAlchemy>=2.0,<3.0",
  "Flask>=3.1,<4.0",
  "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
  "alembic>=1.14,<2.0",
  "pre-commit>=4.0,<5.0",
  "pytest>=8.0,<9.0",
  "ruff>=0.8,<1.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
